\name{RepDbLite.Mmusculus.2103}
\docType{package}

\alias{RepDbLite.Mmusculus.2103-package}
\alias{RepDbLite.Mmusculus.2103}
\alias{RepDbLite.Mmusculus.2103}


\title{RepBase-based annotation package}

\description{
  This package loads an SQL connection to a database containing
  annotations from Ensembl. For examples and help on functions see the
  help pages from the \code{ensembldb} package!
}

\note{
  This data package was made from resources at RepBase on
  Wed Apr 20 11:13:10 2016 and based on the RepBase21_03
}

\author{.}



\examples{
## load the library
##library(RepDbLite.Mmusculus.2103)
## list the contents that are loaded into memory
ls('package:RepDbLite.Mmusculus.2103')
## show the db object that is loaded by calling it's name
RepDbLite.Mmusculus.2103

## for more examples see the ensembldb package.


}

\keyword{package}
\keyword{data}
